requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
downloads==1.0.0
freeze==3.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
